pyproject.toml
[tool.poetry]
name = "meme-bot"
version = "0.1.0"
description = ""
authors = ["ropc <itslikeroar2412@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.2.4"
pillow = "^6.1"
pygtrie = "^2.3"
pydantic = "^1.1"

[tool.poetry.dev-dependencies]
ipython = "^7.7"
rope = "^0.14.0"
ptvsd = "^4.3.2"
mypy = "^0.740"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
